Removed properties from class outside of scope
I removed the “studentCount” properties from your student class, only
because the class itself focuses on a single person, so any object of
that class should just deal with a single person and having a total
count of all students in every student object seemed to be out of scope
of the class.  It would probably be best to create a class to handle
all the stuff that dials with all the students, such as the student
count.  So you’d have the student class which deals with individual
students and the maybe a “classroom” class that has all the properties
for multiple students.  And in your student constructor, you can call a
member function of the “classroom” class that increments the count of
student count in the classrom class.  As of now, I just made a variable
inside the printList to handle counting students.